requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
jinja2>=3.1.0
huggingface_hub>=0.23.0
pyyaml>=6.0.0
python-multipart>=0.0.6
